chore(wip): move TestHttpServer into test/patched/OpenDraftTestServer.ts
