+++ test/modules/fetch/intercept/fetch.clone.browser.test.ts
@@ -1,6 +1,6 @@
 import * as path from 'path'
 import { pageWith } from 'page-with'
-import { HttpServer } from '../../../TestHttpServer'
+import { HttpServer } from '../../../patched/OpenDraftTestServer'
 
 declare namespace window {
   export const fetchData: (url: RequestInfo) => Promise<void>
