+++ test/modules/http/compliance/http-res-read-multiple-times.test.ts
@@ -4,7 +4,7 @@
  * @see https://github.com/mswjs/interceptors/issues/161
  */
 import http, { IncomingMessage } from 'http'
-import { HttpServer } from '../../../TestHttpServer'
+import { HttpServer } from '../../../patched/OpenDraftTestServer'
 import { HttpRequestEventMap, IsomorphicResponse } from '../../../../src'
 import { ClientRequestInterceptor } from '../../../../src/interceptors/ClientRequest'
