+++ test/modules/http/compliance/http-res-set-encoding.test.ts
@@ -2,7 +2,7 @@
  * @jest-environment node
  */
 import * as http from 'http'
-import { HttpServer } from '../../../TestHttpServer'
+import { HttpServer } from '../../../patched/OpenDraftTestServer'
 import { ClientRequestInterceptor } from '../../../../src/interceptors/ClientRequest'
 
 const httpServer = new HttpServer((app) => {
