+++ test/modules/http/intercept/http.request.test.ts
@@ -3,7 +3,7 @@
  */
 import * as http from 'http'
 import { RequestHandler } from 'express-serve-static-core'
-import { HttpServer } from '../../../TestHttpServer'
+import { HttpServer } from '../../../patched/OpenDraftTestServer'
 import { anyUuid, headersContaining } from '../../../jest.expect'
 import { waitForClientRequest } from '../../../helpers'
 import { ClientRequestInterceptor } from '../../../../src/interceptors/ClientRequest'
