+++ test/modules/http/intercept/https.request.test.ts
@@ -3,7 +3,7 @@
  */
 import * as https from 'https'
 import { RequestHandler } from 'express'
-import { HttpServer, httpsAgent } from '../../../TestHttpServer'
+import { HttpServer, httpsAgent } from '../../../patched/OpenDraftTestServer'
 import { waitForClientRequest } from '../../../helpers'
 import { anyUuid, headersContaining } from '../../../jest.expect'
 import { ClientRequestInterceptor } from '../../../../src/interceptors/ClientRequest'
