+++ test/modules/http/regressions/http-socket-timeout.ts
@@ -5,7 +5,7 @@
  * due to the unterminated socket.
  */
 import * as http from 'http'
-import { HttpServer } from '../../../TestHttpServer'
+import { HttpServer } from '../../../patched/OpenDraftTestServer'
 import { ClientRequestInterceptor } from '../../../../src/interceptors/ClientRequest'
 
 jest.setTimeout(5000)
